Workaround configuration sorting
Workaround broken sorting of configurations in configuration groups. The built-in extbase object storage mapper does not respect the order in which the configuration ids are saved in the database, but sorts them by uid which lead to inconclusive results in the calculated time table. This might be related to this bug report: https://forge.typo3.org/issues/27659 (german article: http://blog.teamgeist-medien.de/2014/12/extbase-sortierung-von-child-objects-objectstorages-irre.html)