[FEATURE] Sort events with end_date to the end of search result
New "sortBy" option "withrangelast" in the plugin settings.

Events with a given end_date and a start_date that is before the start date of the output are always displayed first.
With this sort option, these events are sorted according to the end date and only appear in the list after all events without an end date.
Useful if you want to display more current dates at the beginning of the list. Especially if many events have a start date before the start date of the output and an end date far in the future. Otherwise, these are always displayed at the beginning of the list and you have to browse to see the current events.