Do not convert moment.js dates to UTC
When dates are parsed with the `moment.utc` function,
the resulting time point is UTC and will be displayed as such
when it is formatted. This means that the user will only see
UTC times, which is not very helpful. `moment.parseZone` keeps
the time zone specified on the timestamp.