Fix timezone of dates/times in ical event adapters
The dates and times have been calculated with the datetime objects
returned by the iCalendar parser. These had timezones set by the
corresponding VEvent entries. This caused them to be formatted in these
timezones and stored as local timezone in the db.
This fix converts all dates to the local timezone of the server.

Example (assuming +06:00 timezone):
ICal-Date           Before (wrong)      Now
20180614T114000Z    2018-06-14 42000    2018-06-14 63600
20031214T220000Z    2003-12-14 79200    2003-12-15 14400 (next day)
- also applies for events with timezones (e.g. TZID=Europe/Berlin)