Add service to create configuration from iCalEvent
The hydration of configurations for iCalender events is now handled in
a sperate service. This allows the reuse in custom events models.
Additionally it can now parse simple recurring events. It  supports
RRULE with the FREQ, INTERVAL, UNTIL and COUNT parts.

In previous versions imported configurations will not get updated.
This is due the new use of "importId" in configurations. This can lead
to duplicate / outdated configurations in the same event.