Fix datetime conversion in extbase
Native dates in the DB are stored / converted as UTC dates in extbase
conflicting with the Typo3 documentation. This is however fixed in
Typo3 11 (https://review.typo3.org/c/Packages/TYPO3.CMS/+/64053).
The added converter inside the models implement/"backport" this
functionality.

The helper functions can be removed in Typo3 >= 11.