Optimize timeFrameConstraints
The time frame constraint creation now uses DateTime objects instead of
timestamps. The constraints check if the startdate is before the range
end and vice versa. This results in shorter queries.
Furthermore the date and datetime methodes are combined and are
accepting open start / end conditions.

Additionally the signal "addTimeFrameConstraints" is now replaced
by an event, however is still supported by an compatibility signal
dispatcher.