Remove restrictions in cal migration Upgrade Wizard
Remove all restrictions except for deleted in getCalendarizeEventUid().
This is already done in the same way in the other functions in the
cal migration upgrade wizard.

This fixing a problem with the event uid sometimes being returned as 0,
for example for hidden records which resulted in errors in the conversions,
for example in sys_category_record_mm.

Resolves: #496