Handle cal migration with sys_category
In newer versions of cal, the categories were already migrated to sys_category.
In this case, the migration must not be performed and the migration of the
event-to-category relations must be handled differently.

Now, the wizard checks if sys_category_record_mm contains references for cal
and handles the migration differently.

Resolves: #494