Show page title in list of pids in BE module
This change will show the same text for each page in the list of
pids as is already used for the "create new" buttons.

Before, only the pid was shown, now the title is displayed as well.

In order to not fetch the page title twice, we drop the pageTitle
ViewHelper and use the page title as already fetched in the Controller
action.

Resolves: #519