Add slug generation in the indexing process
The indexing process generates a slug for each index.
The slug consists of two parts.
The first part is a base slug for each original record, which is
determined by the SpeakingUrlInterface or a fallback methode.
The next part is a date suffix for each index.
Directly before insertion the slug is made unique.

Related #540