Change section output of cleanup command
Output a title at the beginning.

Rearrange the section output so that a section is displayed for
each reindex and the finding events and cleaning up get a section
each.

Since the first reindex may run quite long, it might be beneficial to
have the section index for the first reindex too.

Resolves: #720

Is regression for

- PR: #627
- issue: #626