[BUGFIX] make order by deterministic
In the moment the default ordering for a list view is
startDate ascending and startTime ascending. If several
events have the exact same start date and time the
order by is none deterministic. This is normal and
expected behavior of MySQL. Problem is that MySQL
servers might produce different paginated lists. Lets
say you display 2 events per page and have 4 events.
The 2nd and 3rd event have the exact same start date
and time. In my case the 2nd event is displayed again
as 3rd event on page 2 of the pagination. The only
way to get around this is to make the order by
deterministic by adding the uid field.

fixes #766