Fix plugin action arguments
Since the calendar plugin got split up into multiple plugins, each of
the plugins has its own action argument signature
(e.g. tx_calendarize_listdetail instead of tx_calendarize_calendar).
This complicates the link generation in the ViewHelpers and is not
backwards compatibility.
Instead of this, we explicitly set the namespace of the plugins.