[FEATURE] Add possibility to configure ICS links
There is now a new checkbox in Flexform to configure whether to
hide ICS links.

Currently, the links are only displayed in Detail.html, but
the flexform field is added to other plugins as well.

In the Fluid Detail.html the ICS link is only set if
settings.hideIcalLink is not set to 1 (so it is still
displayed if setting was saved without new setting).

Resolves: #841